requirements.txt
ibm-watson>=8.0.0
ibm-cloud-sdk-core>=3.0.0
python-dotenv>=1.0.0
